DCMAW-5594: Finer file structure
