+++ deploy/scripts/src/mobile/utils/env.ts
@@ -1,0 +1,5 @@
+export const env = {
+  testClientExecuteUrl: __ENV.MOBILE_TEST_CLIENT_EXECUTE_URL,
+  backendUrl: __ENV.MOBILE_BACKEND_URL,
+  frontendUrl: __ENV.MOBILE_FRONTEND_URL
+}
